feat: add support for both local instances and remote clusters (#87)
**Description**

This PR reworks modusgraph to allow the creation of applications to use
both the internal "engine"-based system as well as traditional Dgraph
clusters.

I have kept the old API in place for backwards compatibility.

There are still a number of TODOs including an update of the CHANGELOG,
but wanted to get this up for review ASAP. St


**Checklist**

- [x] Code compiles correctly and linting passes locally
- [ ] For all _code_ changes, an entry added to the `CHANGELOG.md` file
describing and linking to
      this PR
- [x] Tests added for new functionality, or regression tests for bug
fixes added as applicable